Add files via upload
MatLaTex.m is a very simple script intended to enable the user to programmatically include the results of Matlab computations in Latex documents; the focus is on symbolic computations but numerics and figures can also be used. Hence it is similar in purpose to the various “sweave” packages that exist for both Matlab and other languages, e.g. SageMath, R, Python etc. But the emphasis of MatLaTex is on simplicity.